+++ client/tests/client_test.go
@@ -29,12 +29,12 @@ var _ = Describe("Client", func() {
 			APIVersion: "v1",
 		}
 
-		credentials := &credentials.Credentials{
+		sess := session.New(&credentials.Credentials{
 			Username: "Test Username",
 			Password: "Test Password",
-		}
+		})
 
-		twilioClient := client.New(session.New(credentials), apiClientConfig)
+		twilioClient := client.New(sess, apiClientConfig)
 
 		httpmock.ActivateNonDefault(twilioClient.GetRestyClient().GetClient())
 		defer httpmock.DeactivateAndReset()
@@ -219,7 +219,7 @@ var _ = Describe("Client", func() {
 			httpmock.RegisterResponder("DELETE", "https://test-2.twilio.com/v2/test", httpmock.NewStringResponder(200, ""))
 
 			op := client.Operation{
-				OverrideBaseURL: utils.String(client.CreateBaseURL("test-2", "v2")),
+				OverrideBaseURL: utils.String(client.CreateBaseURL("test-2", "v2", nil, nil)),
 				Method:          http.MethodDelete,
 				URI:             "/test",
 			}
@@ -231,6 +231,119 @@ var _ = Describe("Client", func() {
 			})
 		})
 	})
+
+	Describe("When a DELETE request is made with region and edge configuration", func() {
+		httpmock.RegisterResponder("DELETE", "https://test.dublin.ie1.twilio.com/v1/request", httpmock.NewStringResponder(200, ""))
+
+		apiClientConfig := &client.APIClientConfig{
+			Beta:         false,
+			DebugEnabled: false,
+			RetryConfig: client.RetryConfig{
+				Attempts: 0,
+				WaitTime: 0,
+			},
+			SubDomain:  "test",
+			APIVersion: "v1",
+			Edge:       utils.String("dublin"),
+			Region:     utils.String("ie1"),
+		}
+
+		sess := session.New(&credentials.Credentials{
+			Username: "Test Username",
+			Password: "Test Password",
+		})
+
+		twilioClient := client.New(sess, apiClientConfig)
+
+		httpmock.ActivateNonDefault(twilioClient.GetRestyClient().GetClient())
+		defer httpmock.DeactivateAndReset()
+
+		op := client.Operation{
+			Method: http.MethodDelete,
+			URI:    "/request",
+		}
+
+		err := twilioClient.Send(context.Background(), op, nil, nil)
+
+		It("Then no error should be returned", func() {
+			Expect(err).To(BeNil())
+		})
+	})
+
+	Describe("When a DELETE request is made with region and default edge configuration", func() {
+		httpmock.RegisterResponder("DELETE", "https://test.dublin.us1.twilio.com/v1/request", httpmock.NewStringResponder(200, ""))
+
+		apiClientConfig := &client.APIClientConfig{
+			Beta:         false,
+			DebugEnabled: false,
+			RetryConfig: client.RetryConfig{
+				Attempts: 0,
+				WaitTime: 0,
+			},
+			SubDomain:  "test",
+			APIVersion: "v1",
+			Edge:       utils.String("dublin"),
+		}
+
+		sess := session.New(&credentials.Credentials{
+			Username: "Test Username",
+			Password: "Test Password",
+		})
+
+		twilioClient := client.New(sess, apiClientConfig)
+
+		httpmock.ActivateNonDefault(twilioClient.GetRestyClient().GetClient())
+		defer httpmock.DeactivateAndReset()
+
+		op := client.Operation{
+			Method: http.MethodDelete,
+			URI:    "/request",
+		}
+
+		err := twilioClient.Send(context.Background(), op, nil, nil)
+
+		It("Then no error should be returned", func() {
+			Expect(err).To(BeNil())
+		})
+	})
+
+	Describe("When a DELETE request is made with edge configuration", func() {
+		httpmock.RegisterResponder("DELETE", "https://test.au1.twilio.com/v1/request", httpmock.NewStringResponder(200, ""))
+
+		apiClientConfig := &client.APIClientConfig{
+			Beta:         false,
+			DebugEnabled: false,
+			RetryConfig: client.RetryConfig{
+				Attempts: 0,
+				WaitTime: 0,
+			},
+			SubDomain:  "test",
+			APIVersion: "v1",
+			Region:     utils.String("au1"),
+		}
+
+		sess := session.New(&credentials.Credentials{
+			Username: "Test Username",
+			Password: "Test Password",
+		})
+
+		twilioClient := client.New(sess, apiClientConfig)
+
+		httpmock.ActivateNonDefault(twilioClient.GetRestyClient().GetClient())
+		defer httpmock.DeactivateAndReset()
+
+		op := client.Operation{
+			Method: http.MethodDelete,
+			URI:    "/request",
+		}
+
+		err := twilioClient.Send(context.Background(), op, nil, nil)
+
+		It("Then no error should be returned", func() {
+			Expect(err).To(BeNil())
+		})
+	})
+
 })
 
 type TestStructInput struct {